# swagger/swagger.json
{
    "schemes": [
        "http"
    ],
    "swagger": "2.0",
    "info": {
        "description": "This is a Thermy backend API.",
        "title": "Thermy API",
        "contact": {
            "name": "API Support",
            "email": "andreysapozhkov535@gmail.com"
        },
        "license": {
            "name": "MIT",
            "url": "https://mit-license.org/"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api/v1",
    "paths": {
...
        "/admin/elements/all": {
            "get": {
                "description": "return all elements of structural models in the given text markup layer",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "student"
                ],
                "summary": "Show all elements of structural models in the given text markup layer.",
                "parameters": [
                    {
                        "type": "string",
                        "description": "User authentication token",
                        "name": "token",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Text markup layer",
                        "name": "layer",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_Inspirate789_Thermy-backend_internal_domain_interfaces.OutputModelsDTO"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
...
        "/admin/layers/all": {
            "get": {
                "description": "return all text markup layers",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "student"
                ],
                "summary": "Show all text markup layers.",
                "parameters": [
                    {
                        "type": "string",
                        "description": "User authentication token",
                        "name": "token",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_Inspirate789_Thermy-backend_internal_domain_interfaces.LayersDTO"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
...
        "/admin/models/all": {
            "get": {
                "description": "return all structural models in the given text markup layer",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "student"
                ],
                "summary": "Show all structural models in the given text markup layer.",
                "parameters": [
                    {
                        "type": "string",
                        "description": "User authentication token",
                        "name": "token",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Text markup layer",
                        "name": "layer",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_Inspirate789_Thermy-backend_internal_domain_interfaces.OutputModelsDTO"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/admin/properties": {
            "post": {
                "description": "add new unit properties",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "student"
                ],
                "summary": "Add new unit properties.",
                "parameters": [
                    {
                        "type": "string",
                        "description": "User authentication token",
                        "name": "token",
                        "in": "query",
                        "required": true
                    },
                    {
                        "description": "Unit property names",
                        "name": "propertyNames",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_Inspirate789_Thermy-backend_internal_domain_interfaces.PropertyNamesDTO"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_Inspirate789_Thermy-backend_internal_domain_interfaces.PropertiesIdDTO"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/admin/properties/all": {
            "put": {
                "description": "return all unit properties",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "student"
                ],
                "summary": "Show all unit properties.",
                "parameters": [
                    {
                        "type": "string",
                        "description": "User authentication token",
                        "name": "token",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_Inspirate789_Thermy-backend_internal_domain_interfaces.OutputModelsDTO"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/admin/properties/unit": {
            "put": {
                "description": "return all properties for the given unit in the given text markup layer",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "student"
                ],
                "summary": "Show all properties for the given unit in the given text markup layer.",
                "parameters": [
                    {
                        "type": "string",
                        "description": "User authentication token",
                        "name": "token",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Text markup layer",
                        "name": "layer",
                        "in": "query",
                        "required": true
                    },
                    {
                        "description": "Unit data according to which the search will be performed",
                        "name": "propertiesID",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_Inspirate789_Thermy-backend_internal_domain_interfaces.SearchUnitDTO"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_Inspirate789_Thermy-backend_internal_domain_interfaces.OutputPropertiesDTO"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
...
        "/admin/units": {
            "post": {
                "description": "add new units in the given text markup layer",
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "student"
                ],
                "summary": "Add new units in the given text markup layer.",
                "parameters": [
                    {
                        "type": "string",
                        "description": "User authentication token",
                        "name": "token",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Text markup layer",
                        "name": "layer",
                        "in": "query",
                        "required": true
                    },
                    {
                        "description": "Information about stored units",
                        "name": "unitsDTO",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_Inspirate789_Thermy-backend_internal_domain_interfaces.SaveUnitsDTO"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            },
            "patch": {
                "description": "update existing units in the given text markup layer",
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "student"
                ],
                "summary": "Update existing units in the given text markup layer.",
                "parameters": [
                    {
                        "type": "string",
                        "description": "User authentication token",
                        "name": "token",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Text markup layer",
                        "name": "layer",
                        "in": "query",
                        "required": true
                    },
                    {
                        "description": "Information about updated units",
                        "name": "unitsDTO",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_Inspirate789_Thermy-backend_internal_domain_interfaces.UpdateUnitsDTO"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/admin/units/all": {
            "put": {
                "description": "return all units in the given text markup layer",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "student"
                ],
                "summary": "Show all units in the given text markup layer.",
                "parameters": [
                    {
                        "type": "string",
                        "description": "User authentication token",
                        "name": "token",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Text markup layer",
                        "name": "layer",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_Inspirate789_Thermy-backend_internal_domain_interfaces.OutputUnitsDTO"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
...
        "/educator/elements/all": {
            "get": {
                "description": "return all elements of structural models in the given text markup layer",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "student"
                ],
                "summary": "Show all elements of structural models in the given text markup layer.",
                "parameters": [
                    {
                        "type": "string",
                        "description": "User authentication token",
                        "name": "token",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Text markup layer",
                        "name": "layer",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_Inspirate789_Thermy-backend_internal_domain_interfaces.OutputModelsDTO"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "github_com_Inspirate789_Thermy-backend_internal_domain_interfaces.LayersDTO": {
            "type": "object",
            "properties": {
                "layers": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                }
            }
        },
...
        "github_com_Inspirate789_Thermy-backend_internal_domain_interfaces.OutputModelDTO": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                }
            }
        },
        "github_com_Inspirate789_Thermy-backend_internal_domain_interfaces.OutputModelsDTO": {
            "type": "object",
            "properties": {
                "models": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/github_com_Inspirate789_Thermy-backend_internal_domain_interfaces.OutputModelDTO"
                    }
                }
            }
        },
...
        "github_com_Inspirate789_Thermy-backend_internal_domain_interfaces.PropertiesIdDTO": {
            "type": "object",
            "properties": {
                "properties_id": {
                    "type": "array",
                    "items": {
                        "type": "integer"
                    }
                }
            }
        },
        "github_com_Inspirate789_Thermy-backend_internal_domain_interfaces.PropertyNamesDTO": {
            "type": "object",
            "properties": {
                "property_names": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                }
            }
        },
...
        "github_com_Inspirate789_Thermy-backend_internal_domain_interfaces.SearchUnitDTO": {
            "type": "object",
            "properties": {
                "lang": {
                    "type": "string"
                },
                "text": {
                    "type": "string"
                }
            }
        },
...
    }
}